Continue radio continue impl, fix incorrect artist
Some songs seem to display different artists under different circumstances but don't have a consistent share endpoint like the previous artist duplication issue.

For example:
https://music.youtube.com/watch?v=2CwBFr-Eoxg

It has the correct artist 'ゆうゆ' on YouTube music which is fetched using the /next endpoint.

When viewed on YouTube or through the /player endpoint, it has a different artist 'crownrecord' (publisher?).

Replaced the /player endpoint with the /next endpoint. This also fixes the artist duplication issue, so redundant code for that was removed.